feat(onboarding): consolidate welcome step into agent setup wizard
- Merge theme picker and telemetry toggle into wizard SelectionStep gated by isFirstRun prop
- First-run skips health step for 3-screen onboarding (selection → install → complete)
- Telemetry IPC fires on continue, skip, and dismiss (X/Escape) via onBeforeClose
- Remove standalone WelcomeStep, OnboardingProgressIndicator, and their tests
- Simplify OnboardingFlow to single-step orchestrator
- Add legacy 'welcome' step resume mapping to agentSetup
- Update OnboardingFlow tests for new single-step flow
